+++ juniper/CHANGELOG.md
@@ -27,6 +27,8 @@ See [#569](https://github.com/graphql-rust/juniper/pull/569).
 
 See [#618](https://github.com/graphql-rust/juniper/pull/618).
 
+- Derive macro `GraphQLEnum` supports custom context (see [#621](https://github.com/graphql-rust/juniper/pull/621))
+
 ## Breaking Changes
 
 - `juniper::graphiql` has moved to `juniper::http::graphiql`